feat : 컨슈머 코드 분리
